Fixed bug #001 (bug counting began 25th August, 2018)
Bug Description - Voter tokens are users without password, they're used
to permit voting on protected ballots. They also have access to views
like other fully registered/logged in users.

Fix - restricted the views that required login with a check on if the
request.user has a usable password or not.